refactor: move scoreRelic into relicScorer
